move formatDisplayedTime to utils
