+++ editor/core/utils/index.ts
@@ -42,3 +42,4 @@ export * from "./fade";
 export * from "./frame";
 export * from "./led";
 export * from "./genJson";
+export * from "./timeFormat"
